+++ packages/perseus/src/renderer.tsx
@@ -305,7 +305,9 @@ class Renderer
             );
         }
 
-        this.props.apiOptions?.answerableCallback?.(this._isAnswerable());
+        if (this.props.userInput) {
+            this.props.apiOptions?.answerableCallback?.(this._isAnswerable());
+        }
     }
 
     UNSAFE_componentWillReceiveProps(nextProps: Props) {