+++ packages/keypad-context/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/keypad-context
 
+## 3.2.50
+
+### Patch Changes
+
+-   Updated dependencies [[`48afa54cfa`](https://github.com/Khan/perseus/commit/48afa54cfad3237bbaaa852f554e3018a05c2920)]:
+    -   @khanacademy/perseus-core@26.1.1
+
 ## 3.2.49
 
 ### Patch Changes